+++ src/utils/WindowSegmentHelpers.ts
